maintain: rename cache to redis
