+++ internal/server/server.go
@@ -23,9 +23,9 @@ import (
 	"github.com/infrahq/infra/internal/ginutil"
 	"github.com/infrahq/infra/internal/logging"
 	"github.com/infrahq/infra/internal/repeat"
-	"github.com/infrahq/infra/internal/server/cache"
 	"github.com/infrahq/infra/internal/server/data"
 	"github.com/infrahq/infra/internal/server/email"
+	"github.com/infrahq/infra/internal/server/redis"
 	"github.com/infrahq/infra/metrics"
 )
 
@@ -49,9 +49,8 @@ type Options struct {
 	SessionDuration          time.Duration
 	SessionExtensionDeadline time.Duration
 
-	// Cache contains configuration options to the cache server. Currently
-	// only Redis configuration is supported.
-	Cache cache.Options
+	// Redis contains configuration options to the cache server.
+	Redis redis.Options
 
 	DBEncryptionKey         string
 	DBEncryptionKeyProvider string
@@ -113,7 +112,7 @@ type TLSOptions struct {
 type Server struct {
 	options         Options
 	db              *data.DB
-	cache           *cache.Cache
+	redis           *redis.Redis
 	tel             *Telemetry
 	secrets         map[string]secrets.SecretStorage
 	keys            map[string]secrets.SymmetricKeyProvider
@@ -172,8 +171,8 @@ func New(options Options) (*Server, error) {
 	server.db = db
 	server.metricsRegistry = setupMetrics(server.db)
 
-	if cache := cache.NewCache(options.Cache); cache != nil {
-		server.cache = cache
+	if redis := redis.NewRedis(options.Redis); redis != nil {
+		server.redis = redis
 	}
 
 	if options.EnableTelemetry {
